[SSD-READ] fullTest 부분 일부 수정했습니다.

+++ SSD/src/test/java/ssd/ssdFullTest.java
@@ -2,30 +2,75 @@ package ssd;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import read.ReadCore;
 import read.ReadModule;
+import write.WriteCore;
 import write.WriteModule;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
+import java.io.*;
 
 import static cores.SSDConstraint.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.fail;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ssdFullTest {
+
+    public static final String SAMPLE_VALUE = "0x1289CDEF";
+    public static final int EXITST_VALUE_LBA = 20;
+    public static final int NULL_VALUE_LBA = 10;
+
+    private WriteCore writeCore;
+    private ReadCore readCore;
 
     private final String VALID_VALUE = "0x12341234";
 
     private WriteModule writeModule;
     private ReadModule readModule;
 
+
     @BeforeEach
     void setUp() {
+        writeCore = new WriteModule();
+        readCore = new ReadModule();
+
         this.writeModule = new WriteModule();
         this.readModule = new ReadModule();
+    }
+
+    @Test
+    void 입력한_주소_Value_출력() throws IOException {
+        writeCore.write(EXITST_VALUE_LBA, SAMPLE_VALUE);
+        readCore.read(EXITST_VALUE_LBA);
+
+        assertEquals(SAMPLE_VALUE, getReadResult());
+    }
+
+    @Test
+    void 입력하지않은_주소_Value_출력() throws IOException {
+        writeCore.write(EXITST_VALUE_LBA, SAMPLE_VALUE);
+        readCore.read(NULL_VALUE_LBA);
+
+        assertEquals(DEFAULT_VALUE, getReadResult());
+
+    }
+
+    @Test
+    void write하지않고_read() throws IOException {
+        readCore.read(NULL_VALUE_LBA);
+
+        assertEquals(DEFAULT_VALUE, getReadResult());
+
+    }
+
+    @Test
+    void nand_파일_미존재() throws IOException {
+        deleteNandFile();
+
+        readCore.read(NULL_VALUE_LBA);
+
+        assertEquals(DEFAULT_VALUE, getReadResult());
+
     }
 
     @Test
@@ -117,9 +162,21 @@ public class ssdFullTest {
         assertThat(getReadResult()).isEqualTo("0x00000000");
     }
 
-    private boolean deleteNand() {
+    private static String getReadResult() {
+        try {
+            return new BufferedReader(new FileReader(
+                    (FILE_ABSOLUTE_LOCATION + RESULT_FILENAME)))
+                    .readLine();
+        } catch (IOException e) {
+        }
+        return "";
+    }
+
+    private static void deleteNandFile() {
         File file = new File(FILE_ABSOLUTE_LOCATION + NAND_FILENAME);
-        return file.delete();
+        if (file.exists()) {
+            file.delete();
+        }
     }
 
     private void writeDataToAddressAndRead(int address, String inputValue) {
@@ -127,14 +184,9 @@ public class ssdFullTest {
         this.readModule.read(address);
     }
 
-    private String getReadResult() {
-        String result = null;
-        try {
-            BufferedReader bufferedReader = new BufferedReader(new FileReader(FILE_ABSOLUTE_LOCATION + RESULT_FILENAME));
-            result = bufferedReader.readLine();
-            bufferedReader.close();
-        } catch (IOException ignored) {
-        }
-        return result;
+    private boolean deleteNand() {
+        File file = new File(FILE_ABSOLUTE_LOCATION + NAND_FILENAME);
+        return file.delete();
     }
+
 }
